+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added microgrid photovoltaic actions to myems-api
 - added microgrid windturbine to system database
 - added microgrid windturbine actions to myems-api
+- added get all associated objects of microgrid to myems-api
 ### Changed
 
 ### Fixed